feat:update room works
