+++ client/src/components/room/RoomModal.tsx
@@ -16,7 +16,7 @@ import {
   Check,
   Sparkles,
 } from "lucide-react";
-import { roomApi } from "@/lib/api/room";
+import { roomApi } from "@/lib/api/roomApi";
 import { Button } from "@/components/ui/Button";
 import { Input } from "@/components/ui/Input";
 import { cn } from "@/lib/utils";
