+++ server/src/utils/schema-validation/validation.pipe.ts
@@ -3,20 +3,29 @@ import { ZodError, ZodSchema } from 'zod';
 
 @Injectable()
 export class ZodPipe implements PipeTransform<any> {
-  constructor(private schema: ZodSchema) {}
+  constructor(
+    private schema: ZodSchema,
+    private options?: { forbidEmpty?: boolean },
+  ) {}
 
   transform(value: unknown) {
+    if (
+      this.options?.forbidEmpty &&
+      (!value || Object.keys(value).length === 0)
+    ) {
+      throw new BadRequestException('Request body is empty');
+    }
     try {
       const parsedValue = this.schema.parse(value);
       return parsedValue;
     } catch (err) {
       if (err instanceof ZodError) {
-        const errorMessages: string[] = [];
-        for (let i = 0; i < err.issues.length; i++) {
-          errorMessages.push(err.issues[i].message);
-        }
+        const errorMessages = err.issues.map(
+          (issue) => `${issue.path.join('.')}: ${issue.message}`,
+        );
         throw new BadRequestException(errorMessages);
       }
+      throw err;
     }
   }
 }
